requirements.txt
dnspython
flask
scikit-learn
pandas
redis
joblib
spacy
fuzzywuzzy
python-Levenshtein
